requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
openai==0.28.1
PyPDF2==3.0.1
pdfplumber==0.9.0
requests==2.31.0
pdfplumber==0.9.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
